% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{ml_brain_region_info}
\alias{ml_brain_region_info}
\alias{ml_brain_graph}
\title{Get information about the mouse brain compartments in the MouseLight project}
\usage{
ml_brain_region_info(...)

ml_brain_graph(...)
}
\arguments{
\item{...}{methods passed to \code{ml_fetch}}
}
\description{
Query MouseLight's GraphQL API to retreive information about all of the brain volumes in the MouseLight project. These
brain regions can be sub-volumes of other brain regions. An \code{igraph} object can be generated to explore these dependencies.
}
\examples{
\dontrun{
## First we need to download all of the neurons
mbr = ml_brain_region_info()

## Leet's have a look at this data
View(ml_brain_region_info)

## Some brain regions are actually a subcompartment of others in this set.
### Hmm, so what would be good is if we could see those correspondences
g = ml_brain_graph()

## Let's plot these i na way we can easily see!
library(igraph)
f = igraph::layout_with_fr(g)
plot(g, layout = f)

## There are a lot of nodes here, we might want to move them about
tkplot(g, layout = f)

}
}
\references{
Winnubst, Johan, Erhan Bas, Tiago A. Ferreira, Zhuhao Wu, Michael N. Economo, Patrick Edson, Ben J. Arthur, et al. 2019. “Reconstruction of 1,000 Projection Neurons Reveals New Cell Types and Organization of Long-Range Connectivity in the Mouse Brain.” bioRxiv. https://doi.org/10.1101/537233.
Economo, Michael N., Nathan G. Clack, Luke D. Lavis, Charles R. Gerfen, Karel Svoboda, Eugene W. Myers, and Jayaram Chandrashekar. 2016. “A Platform for Brain-Wide Imaging and Reconstruction of Individual Neurons.” eLife 5 (January): e10566.
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_read_neurons}}
}
