% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ml_list_neurons}
\alias{ml_list_neurons}
\title{Query available neurons from mouselight site}
\usage{
ml_list_neurons(simplify = TRUE)
}
\arguments{
\item{simplify}{Whether to simplify the query result to a simple data.frame
with one row per tracing (See details).}
}
\value{
data.frame describing the available neurons. See details for the
  implication of the \code{simplify} argument.
}
\description{
Query available neurons from mouselight site
}
\details{
Mouselight neurons typically consist of two separate tracings, one
  for the axon and one for the dendrites. When \code{simplify=TRUE} (the
  default) the output is a \emph{flattened} data frame with one row for every
  tracing. When \code{simplify=FALSE} there is one row per neuron, but the
  data.frame contains nested data frames with per tracing information.

  The identifier which you will need to download a neuron with
  \code{\link{read.neurons.mouselight}} is stored in column
}
\examples{
\donttest{
ndf=ml_list_neurons()
# How many tracings per neurons?
table(table(ndf$neuron.id))
}
}
