% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mouselight_brain_region_info}
\alias{mouselight_brain_region_info}
\alias{mouselight_brain_graph}
\title{Get information about the mouse brain compartments in the MouseLight project}
\usage{
mouselight_brain_region_info(...)

mouselight_brain_graph(...)
}
\arguments{
\item{...}{methods passed to \code{mouselight_fetch}}
}
\description{
Query MouseLight's GraphQL API to retreive information about all of the brain volumes in the MouseLight project. These
brain regions can be sub-volumes of other brain regions. An \code{igraph} object can be generated to explore these dependencies.
}
\examples{
\dontrun{
## First we need to download all of the neurons
mbr = mouselight_brain_region_info()

## Leet's have a look at this data
View(mouselight_brain_region_info)

## Some brain regions are actually a subcompartment of others in this set.
### Hmm, so what would be good is if we could see those correspondences
g = mouselight_brain_graph()

## Let's plot these i na way we can easily see!
library(igraph)
f = igraph::layout_with_fr(g)
plot(g, layout = f)

## There are a lot of nodes here, we might want to move them about
tkplot(g, layout = f)

}
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_read_neurons}}
}
