% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mouselight_api}
\alias{mouselight_api}
\title{Get information about the MouseLight project's GraphQL API}
\usage{
mouselight_api(...)
}
\arguments{
\item{...}{methods passed to \code{mouselight_fetch}}
}
\value{
a named vector of values
}
\description{
See the API version and the number of neurons hosted by the project.
}
\examples{
\dontrun{
## First we need to download all of the neurons
mouselight_api(...)

}
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_read_neurons}}
}
