% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{mouselight_neuron_info}
\alias{mouselight_neuron_info}
\alias{mouselight_neuron_somata}
\title{Query available neurons from MouseLight site}
\usage{
mouselight_neuron_info(simplify = TRUE)

mouselight_neuron_somata()
}
\arguments{
\item{simplify}{whether to simplify the query result to a simple data.frame
with one row per tracing (See details).}
}
\value{
data.frame describing the available neurons. See details for the
  implication of the \code{simplify} argument.
}
\description{
Query tracings for neurons, and/or neuron soma locations, from the MouseLight Neuron Browser.
Each neuron may have multiple tracgings, typically one for its
axon and another for its dendrite.
}
\details{
Mouselight neurons typically consist of two separate tracings, one
  for the axon and one for the dendrites. When \code{simplify=TRUE} (the
  default) the output is a \emph{flattened} data frame with one row for every
  tracing. When \code{simplify=FALSE} there is one row per neuron, but the
  data.frame contains nested data frames with per tracing information.

  The identifier which you will need to download a neuron with
  \code{\link{mouselight_read_neurons}} is stored in column
}
\examples{
\donttest{
ndf=mouselight_neuron_info()
# How many tracings per neurons?
table(table(ndf$neuron.id))

# In what brain regions do neurons have their end points?
br_search = mouselight_neurons_by_brain_region("dendrite")
View(br_search)

}
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_brain_region_info}}
}
