% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meshes.R
\name{mouselight_read_brain}
\alias{mouselight_read_brain}
\title{Read the brain mesh and its brain regions from the MouseLight project}
\usage{
mouselight_read_brain(type = c("outline", "brain_areas"),
  progress = TRUE, Force = FALSE)
}
\arguments{
\item{type}{whether just to retreive the outer brain mesh, or the brain constituent neuropils}

\item{progress}{logical, toggle progress bar}

\item{Force}{logical, whether or not to overwrite .obj files stored in a temporary directory}

\item{...}{methods passed to \code{mouselight_fetch}}
}
\value{
a \code{nat} package \code{\link[nat]{hxsurf}} object, which mimics the Amira surface format, replete with metadata that can be
accessed using \code{$}
}
\description{
Save  the brain mesh and its brain regions from the MouseLight project
}
\examples{
\dontrun{
## First we can quickly just plot the outer mesh for the brain
outline = mouselight_read_brain(type = "outline")
plot3d(outline, col = "pink", alpha = 0.3)

## This is cool, but maybe what we really want are its sub-divisions.
mousebrain = mouselight_read_brain(type = "brain_areas")
clear3d()
plot3d(mousebrain)
### This takes a long time the first time you call this function per session

## What brain regions are on offer?
print(mousebrain$neuropil_full_names)

## Or if we want more information, we can get it like this:
mbr = mouselight_brain_region_info()
View(mouselight_brain_region_info)

## Perhaps we want to plot just the amygdala?
### To do this we can do
amygdala.codes = mousebrain$RegionList[grepl("amygdala",
mousebrain$neuropil_full_names,
ignore.case = TRUE)]
plot3d(outline, col = "pink", alpha = 0.1)
plot3d(subset(mousebrain, amygdala.codes), alpha = 0.5)

}
}
\references{
Winnubst, Johan, Erhan Bas, Tiago A. Ferreira, Zhuhao Wu, Michael N. Economo, Patrick Edson, Ben J. Arthur, et al. 2019. “Reconstruction of 1,000 Projection Neurons Reveals New Cell Types and Organization of Long-Range Connectivity in the Mouse Brain.” bioRxiv. https://doi.org/10.1101/537233.
Economo, Michael N., Nathan G. Clack, Luke D. Lavis, Charles R. Gerfen, Karel Svoboda, Eugene W. Myers, and Jayaram Chandrashekar. 2016. “A Platform for Brain-Wide Imaging and Reconstruction of Individual Neurons.” eLife 5 (January): e10566.
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_brain_region_info}}
}
