% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouselightr-package.R
\docType{package}
\name{mouselightr-package}
\alias{mouselightr}
\alias{mouselightr-package}
\title{mouselightr: Bridge HHMI Janelia MouseLight Database and nat R Package}
\description{
MouseLight provide an interactive 3D browser linked from
    http://mouselight.janelia.org. This package allows neurons from
    that dataset to be downloaded and manipulated offline, providing a
    bridge to the nat (NeuroAnatomy Toolbox) suite of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/mouselightr}
  \item Report bugs at \url{https://github.com/natverse/mouselightr/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Alexander Bates \email{alexander.shakeel.bates@gmail.com} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})
}

}
\keyword{internal}
