% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.mouselight-package.R
\docType{package}
\name{nat.mouselight-package}
\alias{nat.mouselight}
\alias{nat.mouselight-package}
\title{nat.mouselight: Bridge HHMI Janelia MouseLight Database and nat R Package}
\description{
MouseLight provide an interactive 3D browser at
    http://ml-neuronbrowser.janelia.org/. This package allows neurons from
    that dataset to be downloaded and manipulated offline, providing a
    bridge to the nat (NeuroAnatomy Toolbox) suite of R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/mouselightr}
  \item Report bugs at \url{https://github.com/natverse/mouselightr/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

Authors:
\itemize{
  \item Alexander Bates \email{alexander.shakeel.bates@gmail.com} (0000-0002-1195-0445)
}

}
\keyword{internal}
