% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuron-io.R
\name{read.neurons.mouselight}
\alias{read.neurons.mouselight}
\title{Read one or more mouselight tracings into a neuronlist}
\usage{
read.neurons.mouselight(x)
}
\arguments{
\item{x}{ids of mouselight neurons}
}
\value{
A \code{nat::\link{neuronlist}}
}
\description{
Read one or more mouselight tracings into a neuronlist
}
\examples{
\donttest{
ids=c("1ccbc478-88fb-4181-a4de-d0c22ed9738b",
  "3afa5e41-374d-45bc-a546-03f362d93649")
nn=read.neurons.mouselight(ids)
library(nat)
plot3d(nn)
}
}
