% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mouselight_api}
\alias{mouselight_api}
\title{Get information about the MouseLight project's GraphQL API}
\usage{
mouselight_api(...)
}
\arguments{
\item{...}{methods passed to \code{mouselight_fetch}}
}
\value{
a named vector of values
}
\description{
See the API version and the number of neurons hosted by the project.
}
\examples{
\donttest{
mouselight_api()
}
}
\references{
Winnubst, Johan, Erhan Bas, Tiago A. Ferreira, Zhuhao Wu, Michael N. Economo, Patrick Edson, Ben J. Arthur, et al. 2019. “Reconstruction of 1,000 Projection Neurons Reveals New Cell Types and Organization of Long-Range Connectivity in the Mouse Brain.” bioRxiv. https://doi.org/10.1101/537233.
Economo, Michael N., Nathan G. Clack, Luke D. Lavis, Charles R. Gerfen, Karel Svoboda, Eugene W. Myers, and Jayaram Chandrashekar. 2016. “A Platform for Brain-Wide Imaging and Reconstruction of Individual Neurons.” eLife 5 (January): e10566.
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_read_neurons}}
}
