% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{mouselight_neuron_info}
\alias{mouselight_neuron_info}
\alias{mouselight_neuron_somata}
\title{Query available neurons from MouseLight site}
\usage{
mouselight_neuron_info(simplify = TRUE)

mouselight_neuron_somata(soma.info = TRUE)
}
\arguments{
\item{simplify}{whether to simplify the query result to a simple data.frame
with one row per tracing (See details).}

\item{structure}{when searching for the brain regions containing neurons of interest,
the part of the neuron that should be used for the search}
}
\value{
data.frame describing the available neurons. See details for the
  implication of the \code{simplify} argument.
}
\description{
Query available neurons from MouseLight site
}
\examples{
\donttest{
ndf=mouselight_neuron_info()
# How many tracings per neurons?
table(table(ndf$neuron.id))

# In what brain regions do neurons have their end points?
br_search = mouselight_neurons_by_brain_region("dendrite")
View(br_search)

}
}
\seealso{
\code{\link{mouselight_read_brain}}, \code{\link{mouselight_brain_region_info}}
}
